feat(history_pane): Prepend results, keep view at top with new results (zelikos/rollit!49)
Previously, new roll results were added to the bottom of the history list. Theoretically, the view would've scrolled to the bottom as new results were added. However, the view would only ever scroll to the second-to-last result despite attempts to fix this and referencing other apps that do similar "stick to bottom" scrolling behavior.

Now, new roll results are prepended to the top of the history list, so results are displayed in reverse chronological order (e.g. more like browser history, instead of chat apps). When a new result is prepended, the view is kept at the top. If the user has scrolled the view down, the view stays in place if new results are added.