ci(github): Don't specify gnome-nightly repo
Looking over flatpak-github-action again, the container image *should*
have Flathub, Flathub-Beta, and gnome-nightly as available remotes, so
specifying gnome-nightly isn't necessary (and prevents pulling the rust
runtime from Flathub).

I remember having issues with using the Devel manifest with this action
before, but that was also before bumping the version of the action. So,
fingers crossed that it just works now