refactor: Build AboutWindow from appdata (zelikos/rollit!67)
Builds AboutWindow from appdata instead of setting it up manually as before. This reduces maintenance burden by re-using applicable info from the appdata, and also automatically pulls in the release notes from the specified version. 

Doing this also required setting some of the properties in code again, so the data for said properties are assigned to constants before being set to the About window.