feat(tray): Add reset button (zelikos/rollit!86)
Adds a "Reset Tray" button to restore the tray to its default state. A confirmation dialog is presented when the button is clicked; in the future, I might redo this as an "Undo" toast instead, but that's low priority for now.

Also re-organizes the tray-row files into the `widgets` subdirectories, as they make more sense in there than in `dialogs`.

Finally, this switches the runtime version back to `master` because `46beta` apparently still isn't on Flathub yet. We'll switch back once it is available.

Fixes #45