feat: Improve tray item removal (zelikos/rollit!87)
Makes the "delete" button inline to each row
Moves the reset button to the header area where the delete button used to be.

I'm using the `circular` style on the remove button because it matches better with the SpinRow buttons, but `flat` might be more HIG-appropriate...? If so, the style can easily be revisited later.

Fixes #47